requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
django-cors-headers==4.7.0
python-decouple==3.8
celery==5.5.3
# Python 3.10.12